fix: コードレビュー指摘のバグ7件とflakyテストを修正
レビューで CONFIRMED/PLAUSIBLE と判定した修正:
- core: テーブル row_count を uint16_t→uint32_t に拡張し、65536行超のラップと
  cell_text_starts 不変条件の破綻を防止
- app/layout: スクロール上限を末尾 spacing_below 抜き
  (LayoutService::GetScrollableContentHeight) に統一し、スクロールバー描画/
  ドラッグ/領域判定/SyncMaxScroll の5箇所を揃えてオーバースクロールを解消
- app: 検索ジャンプ (on_scroll_changed) で SyncTocActive を発火し、TOC の
  アクティブ見出しを追従させる
- io: LoadInt で from_chars の戻り ptr を検証し、"300px" 等の部分パースを拒否
- app: FinishLoadMarkdownFile の未使用 scroll_y デッドストアを削除

テスト改善:
- test_view_stats を MENDO_EXAMPLE_DIR + std::filesystem::path で cwd 非依存化
  (常時SKIP解消、日本語UTF-8パス対応)
- MockTextMeasurer のテーブル計測を実装と内部整合 (border 定数統一、列幅は
  均等割り近似である制約を明記)
- ReducerTest fixture の Theme を value-init し、未初期化フィールド由来の
  順序依存 flaky (TocItemClicked_TailSection) を修正

全2307テスト PASS。

+++ src/core/document_types.h
@@ -96,7 +96,9 @@ struct NodeTableData {
     // 行単位の header フラグ (md4c は TR 内で TH/TD が混在しない契約)
     std::pmr::vector<bool> is_header_row;
 
-    uint16_t row_count = 0;
+    // row は数十万〜数百万行になりうるため uint32_t。col_count は parser 側で uint16 上限に
+    // クランプ済み (65535 列超は非現実的)。両者の積 (cell index) は CellIndex で size_t 昇格。
+    uint32_t row_count = 0;
     uint16_t col_count = 0;
 
     constexpr size_t CellIndex(size_t r, size_t c) const noexcept
